Refactor : scroll component 분리
